[62] Adjust routes, index, clientSchema and clientControllers. Fix mongoDB and docker files
