+++ .env_db
@@ -1,0 +1,4 @@
+DB_USER=api_user
+DB_PASS=api_password
+DB_NAME=clients_database
+DB_HOST=db_clients
