+++ Dockerfile
@@ -1,0 +1,7 @@
+FROM node:15
+
+COPY . /code
+
+WORKDIR /code
+
+CMD ["sh","-c", "yarn install && yarn start"]
