+++ docker-compose.yml
@@ -1,0 +1,41 @@
+version: '3.8'
+
+services:
+  api_clients:
+    container_name: backend_clients
+    build: .
+    ports:
+      - 3002:3002
+    environment: 
+      PORT: 3002
+    env_file: .env_db
+    volumes:
+      - ./:/code
+      - node_modules:/code/node_modules
+    networks: 
+      - backend_clients
+    depends_on:
+      - db_clients
+
+  db_clients:
+    container_name: db_clients
+    image: mongo:4.4
+    restart: always
+    environment:
+      MONGO_INITDB_ROOT_USERNAME: root
+      MONGO_INITDB_ROOT_PASSWORD: password
+      MONGO_INITDB_DATABASE: root-db-client
+    env_file: .env_db
+    volumes:
+      - ./db/mongo-init.sh:/docker-entrypoint-initdb.d/mongo-init.sh:ro
+      - ./db/.dbshell:/home/mongodb/.dbshell
+      - mongodb:/data/db
+    networks: 
+      - backend_clients
+
+volumes: 
+  node_modules:
+  mongodb:
+
+networks: 
+  backend_clients:
