+++ package.json
@@ -6,11 +6,10 @@
   "author": "",
   "license": "MIT",
   "scripts": {
-    "start": "nodemon index.js"
+    "start": "nodemon src/index.js"
   },
   "dependencies": {
     "cors": "^2.8.5",
-    "dotenv-safe": "^8.2.0",
     "express": "^4.17.1",
     "mongoose": "^5.11.19"
   },
